Added Whatsapp send message logic | Added log on whatsapp webhook

+++ src/database/db.ts
@@ -21,24 +21,29 @@ export default class DB {
         return DB.data[key];
     }
 
-    static getMatching(keyPrefix: string): { [key: string]: any } {
+    static getMatching(keyPrefix: string = "", suffix: string = ""): { [key: string]: any } {
         const data = fs.readFileSync('./src/database/data.json', 'utf-8');
         DB.data = JSON.parse(data);
 
         const matching: { [key: string]: any } = {};
         for (const key in DB.data) {
-            if (key.startsWith(keyPrefix)) {
+            if (key.startsWith(keyPrefix) && key.endsWith(suffix)) {
                 matching[key] = DB.data[key];
             }
         }
         return matching;
     }
 
-    static loadFile(fileName: string): string {
-        return fs.readFileSync(`./src/database/files/${fileName}`, 'utf-8');
+    static loadFile(fileName: string, force: boolean = false): string {
+        const filePath = `./src/database/files/${fileName}`;
+        if (!fs.existsSync(filePath) && force) fs.writeFileSync(filePath, '');
+        
+        return fs.readFileSync(filePath, 'utf-8');
     }
 
-    static saveFile(fileName: string, content: string): void {
+    static saveFile(fileName: string, content: string, force: boolean = false): void {
+        if (!fs.existsSync(`./src/database/files/${fileName}`) && !force) throw new Error(`File ${fileName} does not exist.`);
+        
         fs.writeFileSync(`./src/database/files/${fileName}`, content);
     }
 }
